Moved UI to views package and api files to data package.
