+++ app/build.gradle
@@ -51,7 +51,9 @@ dependencies {
     implementation('com.squareup.retrofit2:retrofit:2.3.0') {
         exclude group: 'com.squareup.okhttp3'
     }
-    implementation 'com.squareup.retrofit2:converter-gson:2.3.0'
+    implementation ('com.squareup.retrofit2:converter-gson:2.3.0') {
+        exclude group: 'com.squareup.okhttp3'
+    }
     // Image rendering.
     implementation 'com.squareup.picasso:picasso:2.5.2'
     // Additional dependencies
