+++ app/src/main/java/xyz/jhughes/laundry/injection/AppComponent.java
@@ -3,9 +3,9 @@ package xyz.jhughes.laundry.injection;
 import javax.inject.Singleton;
 
 import dagger.Component;
-import xyz.jhughes.laundry.activities.LocationActivity;
-import xyz.jhughes.laundry.activities.MachineActivity;
-import xyz.jhughes.laundry.fragments.MachineFragment;
+import xyz.jhughes.laundry.views.activities.LocationActivity;
+import xyz.jhughes.laundry.views.activities.MachineActivity;
+import xyz.jhughes.laundry.views.fragments.MachineFragment;
 
 @Singleton
 @Component(modules = AppModule.class)
