+++ app/src/main/java/xyz/jhughes/laundry/views/adapters/MachineAdapter.java
@@ -1,4 +1,4 @@
-package xyz.jhughes.laundry.adapters;
+package xyz.jhughes.laundry.views.adapters;
 
 import android.content.Context;
 import android.content.DialogInterface;
@@ -24,22 +24,20 @@ import javax.inject.Inject;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
-import xyz.jhughes.laundry.BuildConfig;
 import xyz.jhughes.laundry.R;
 import xyz.jhughes.laundry.SnackbarPostListener;
 import xyz.jhughes.laundry.analytics.AnalyticsHelper;
-import xyz.jhughes.laundry.apiclient.MachineService;
 import xyz.jhughes.laundry.databinding.CardviewMachineBinding;
 import xyz.jhughes.laundry.laundryparser.Constants;
 import xyz.jhughes.laundry.laundryparser.Machine;
 import xyz.jhughes.laundry.laundryparser.MachineStates;
 import xyz.jhughes.laundry.laundryparser.MachineTypes;
-import xyz.jhughes.laundry.apiclient.MachineAPI;
+import xyz.jhughes.laundry.data.MachineAPI;
 import xyz.jhughes.laundry.notificationhelpers.NotificationCreator;
 import xyz.jhughes.laundry.notificationhelpers.OnMachineChangedToInUse;
 import xyz.jhughes.laundry.notificationhelpers.ScreenOrientationLockToggleListener;
 import xyz.jhughes.laundry.runnables.MachineCheckerRunnable;
-import xyz.jhughes.laundry.storage.SharedPrefsHelper;
+import xyz.jhughes.laundry.views.storage.SharedPrefsHelper;
 
 import static xyz.jhughes.laundry.laundryparser.MachineStates.AVAILABLE;
